+++ ckanext/search_tweaks/tests/spellcheck/test_plugin.py
@@ -6,7 +6,7 @@ from bs4 import BeautifulSoup
 import ckan.plugins as p
 import ckan.plugins.toolkit as tk
 from ckan.tests.factories import Dataset
-
+from ckanext.search_tweaks.spellcheck.plugin import CONFIG_SHOW_ONLY_MORE
 
 @pytest.mark.ckan_config("ckan.plugins", "search_tweaks search_tweaks_spellcheck")
 @pytest.mark.usefixtures("with_plugins")
@@ -47,3 +47,16 @@ class TestHelper:
 
         assert helper("pic", 2) is None
         assert helper("pic", 1) == "pick"
+
+    def test_show_only_more_results(self, ckan_config, monkeypatch):
+        Dataset(title="Pick this")
+        Dataset(notes="Pick this")
+        Dataset(title="Pock this")
+        helper = tk.h.spellcheck_did_you_mean
+
+        assert helper("pock", 1) == "pick"
+        assert helper("pick", 2) is None
+
+        monkeypatch.setitem(ckan_config, CONFIG_SHOW_ONLY_MORE, "off")
+        assert helper("pock", 1) == "pick"
+        assert helper("pick", 2) == "pock"